+++ tests/test_video.py
@@ -75,7 +75,7 @@ def imgseq_fixture():
     imgseq_dir = Path(IMGSEQ_DIR)
     imgseq_dir.mkdir(parents=True, exist_ok=True)
     for i, frame in enumerate(cv3.Video(TEST_VID)):
-        cv3.imwrite(imgseq_dir / f'img{i//30:02d}.png', frame)
+        cv3.imwrite(imgseq_dir / 'img{:02d}.png'.format(i//30), frame)
     yield
     shutil.rmtree(IMGSEQ_DIR, ignore_errors=True)
 
@@ -164,11 +164,11 @@ def test_video_extra_kw():
 
 @pytest.fixture()
 def out_path_fixture():
-    Path(OUT_PATH_VID).unlink(missing_ok=True)
-    Path(OUT_PATH_VID_AVI).unlink(missing_ok=True)
+    if Path(OUT_PATH_VID).exists(): Path(OUT_PATH_VID).unlink()
+    if Path(OUT_PATH_VID_AVI).exists(): Path(OUT_PATH_VID_AVI).unlink()
     yield
-    Path(OUT_PATH_VID).unlink(missing_ok=True)
-    Path(OUT_PATH_VID_AVI).unlink(missing_ok=True)
+    if Path(OUT_PATH_VID).exists(): Path(OUT_PATH_VID).unlink()
+    if Path(OUT_PATH_VID_AVI).exists(): Path(OUT_PATH_VID_AVI).unlink()
 
 
 class TestWriterOpenWrite: